business/business-antenna-connector/src/main/java/eu/estcube/gsconnector/commands/radio/SetTranceiveMode.java
package eu.estcube.gsconnector.commands.radio;

import eu.estcube.domain.TelemetryCommand;
import eu.estcube.gsconnector.commands.HamlibCommandBuilder;

public class SetTranceiveMode implements HamlibCommandBuilder {

    public StringBuilder createMessageString(TelemetryCommand command) {
        StringBuilder messageString = new StringBuilder();
        messageString.append("+A ");
        messageString.append(command.getParameter("Transceive"));
        messageString.append("\n");
        return messageString;
    }
}
